pyproject.toml
[project]
name = "market-data-system"
version = "0.1.0"
description = "Market data collection and distribution system"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.103.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.23.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.3",
    "asyncpg>=0.28.0",
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.1",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.10"
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true

[tool.setuptools]
packages = ["core", "api", "db", "fetchers"]

[[tool.mypy.overrides]]
module = ["asyncpg", "asyncpg.*"]
ignore_missing_imports = true